Rework DI loading
